requirements.txt
Flask
Werkzeug
Flask-SQLAlchemy
PyMySQL
python-dotenv
